Fixed compilation warnings on MSVC caused by conversion from pointer to unsigned

+++ source/aarch64/neon/bridge_aarch64_neon.c
@@ -103,8 +103,8 @@ HIDDEN void *rbdBridgeIdenticalWorkerNeon(struct rbdBridgeData *data)
     time = data->batchIdx * V2D;
 
     /* Align, if possible, to vector size */
-    if (((long)&data->reliabilities[time] & (S1D * sizeof(double) - 1)) == 0) {
-        if (((long)&data->reliabilities[time] & (V2D * sizeof(double) - 1)) != 0) {
+    if (((uintptr_t)&data->reliabilities[time] & (S1D * sizeof(double) - 1)) == 0) {
+        if (((uintptr_t)&data->reliabilities[time] & (V2D * sizeof(double) - 1)) != 0) {
             /* Compute reliability of Bridge RBD at current time instant */
             rbdBridgeIdenticalStepS1d(data, time);
             /* Increment current time instant */
