+++ source/aarch64/neon/koon_aarch64_neon.c
@@ -156,8 +156,8 @@ HIDDEN void *rbdKooNIdenticalWorkerNeon(struct rbdKooNIdenticalData *data)
     /* If compute unreliability flag is not set... */
     if (data->bComputeUnreliability == 0) {
         /* Align, if possible, to vector size */
-        if (((long)&data->reliabilities[time] & (S1D * sizeof(double) - 1)) == 0) {
-            if (((long)&data->reliabilities[time] & (V2D * sizeof(double) - 1)) != 0) {
+        if (((uintptr_t)&data->reliabilities[time] & (S1D * sizeof(double) - 1)) == 0) {
+            if (((uintptr_t)&data->reliabilities[time] & (V2D * sizeof(double) - 1)) != 0) {
                 /* Compute reliability of KooN RBD at current time instant from working components */
                 rbdKooNIdenticalSuccessStepS1d(data, time);
                 /* Increment current time instant */
@@ -182,8 +182,8 @@ HIDDEN void *rbdKooNIdenticalWorkerNeon(struct rbdKooNIdenticalData *data)
     }
     else {
         /* Align, if possible, to vector size */
-        if (((long)&data->reliabilities[time] & (S1D * sizeof(double) - 1)) == 0) {
-            if (((long)&data->reliabilities[time] & (V2D * sizeof(double) - 1)) != 0) {
+        if (((uintptr_t)&data->reliabilities[time] & (S1D * sizeof(double) - 1)) == 0) {
+            if (((uintptr_t)&data->reliabilities[time] & (V2D * sizeof(double) - 1)) != 0) {
                 /* Compute reliability of KooN RBD at current time instant from failed components */
                 rbdKooNIdenticalFailStepS1d(data, time);
                 /* Increment current time instant */
