+++ source/amd64/avx512f/koon_amd64_avx512f.c
@@ -189,20 +189,20 @@ HIDDEN void *rbdKooNIdenticalWorkerAvx512f(struct rbdKooNIdenticalData *data)
     /* If compute unreliability flag is not set... */
     if (data->bComputeUnreliability == 0) {
         /* Align, if possible, to vector size */
-        if (((long)&data->reliabilities[time] & (S1D * sizeof(double) - 1)) == 0) {
-            if (((long)&data->reliabilities[time] & (V2D * sizeof(double) - 1)) != 0) {
+        if (((uintptr_t)&data->reliabilities[time] & (S1D * sizeof(double) - 1)) == 0) {
+            if (((uintptr_t)&data->reliabilities[time] & (V2D * sizeof(double) - 1)) != 0) {
                 /* Compute reliability of KooN RBD at current time instant from working components */
                 rbdKooNIdenticalSuccessStepS1d(data, time);
                 /* Increment current time instant */
                 time += S1D;
             }
-            if (((long)&data->reliabilities[time] & (V4D * sizeof(double) - 1)) != 0) {
+            if (((uintptr_t)&data->reliabilities[time] & (V4D * sizeof(double) - 1)) != 0) {
                 /* Compute reliability of KooN RBD at current time instant from working components */
                 rbdKooNIdenticalSuccessStepV2dFma3(data, time);
                 /* Increment current time instant */
                 time += V2D;
             }
-            if (((long)&data->reliabilities[time] & (V8D * sizeof(double) - 1)) != 0) {
+            if (((uintptr_t)&data->reliabilities[time] & (V8D * sizeof(double) - 1)) != 0) {
                 /* Compute reliability of KooN RBD at current time instant from working components */
                 rbdKooNIdenticalSuccessStepV4dFma3(data, time);
                 /* Increment current time instant */
@@ -241,20 +241,20 @@ HIDDEN void *rbdKooNIdenticalWorkerAvx512f(struct rbdKooNIdenticalData *data)
     }
     else {
         /* Align, if possible, to vector size */
-        if (((long)&data->reliabilities[time] & (S1D * sizeof(double) - 1)) == 0) {
-            if (((long)&data->reliabilities[time] & (V2D * sizeof(double) - 1)) != 0) {
+        if (((uintptr_t)&data->reliabilities[time] & (S1D * sizeof(double) - 1)) == 0) {
+            if (((uintptr_t)&data->reliabilities[time] & (V2D * sizeof(double) - 1)) != 0) {
                 /* Compute reliability of KooN RBD at current time instant from failed components */
                 rbdKooNIdenticalFailStepS1d(data, time);
                 /* Increment current time instant */
                 time += S1D;
             }
-            if (((long)&data->reliabilities[time] & (V4D * sizeof(double) - 1)) != 0) {
+            if (((uintptr_t)&data->reliabilities[time] & (V4D * sizeof(double) - 1)) != 0) {
                 /* Compute reliability of KooN RBD at current time instant from failed components */
                 rbdKooNIdenticalFailStepV2dSse2(data, time);
                 /* Increment current time instant */
                 time += V2D;
             }
-            if (((long)&data->reliabilities[time] & (V8D * sizeof(double) - 1)) != 0) {
+            if (((uintptr_t)&data->reliabilities[time] & (V8D * sizeof(double) - 1)) != 0) {
                 /* Compute reliability of KooN RBD at current time instant from failed components */
                 rbdKooNIdenticalFailStepV4dAvx(data, time);
                 /* Increment current time instant */
