+++ source/amd64/fma3/koon_amd64_fma3.c
@@ -120,14 +120,14 @@ HIDDEN void *rbdKooNIdenticalWorkerFma3(struct rbdKooNIdenticalData *data)
     /* If compute unreliability flag is not set... */
     if (data->bComputeUnreliability == 0) {
         /* Align, if possible, to vector size */
-        if (((long)&data->reliabilities[time] & (S1D * sizeof(double) - 1)) == 0) {
-            if (((long)&data->reliabilities[time] & (V2D * sizeof(double) - 1)) != 0) {
+        if (((uintptr_t)&data->reliabilities[time] & (S1D * sizeof(double) - 1)) == 0) {
+            if (((uintptr_t)&data->reliabilities[time] & (V2D * sizeof(double) - 1)) != 0) {
                 /* Compute reliability of KooN RBD at current time instant from working components */
                 rbdKooNIdenticalSuccessStepS1d(data, time);
                 /* Increment current time instant */
                 time += S1D;
             }
-            if (((long)&data->reliabilities[time] & (V4D * sizeof(double) - 1)) != 0) {
+            if (((uintptr_t)&data->reliabilities[time] & (V4D * sizeof(double) - 1)) != 0) {
                 /* Compute reliability of KooN RBD at current time instant from working components */
                 rbdKooNIdenticalSuccessStepV2dFma3(data, time);
                 /* Increment current time instant */
@@ -159,14 +159,14 @@ HIDDEN void *rbdKooNIdenticalWorkerFma3(struct rbdKooNIdenticalData *data)
     }
     else {
         /* Align, if possible, to vector size */
-        if (((long)&data->reliabilities[time] & (S1D * sizeof(double) - 1)) == 0) {
-            if (((long)&data->reliabilities[time] & (V2D * sizeof(double) - 1)) != 0) {
+        if (((uintptr_t)&data->reliabilities[time] & (S1D * sizeof(double) - 1)) == 0) {
+            if (((uintptr_t)&data->reliabilities[time] & (V2D * sizeof(double) - 1)) != 0) {
                 /* Compute reliability of KooN RBD at current time instant from failed components */
                 rbdKooNIdenticalFailStepS1d(data, time);
                 /* Increment current time instant */
                 time += S1D;
             }
-            if (((long)&data->reliabilities[time] & (V4D * sizeof(double) - 1)) != 0) {
+            if (((uintptr_t)&data->reliabilities[time] & (V4D * sizeof(double) - 1)) != 0) {
                 /* Compute reliability of KooN RBD at current time instant from failed components */
                 rbdKooNIdenticalFailStepV2dSse2(data, time);
                 /* Increment current time instant */
